requirements.txt
# =========================================
# requirements.txt - Weather Prediction Pipeline
# Python: 3.10
# Airflow: 2.10.2
# Updated: December 2025
# =========================================

# -----------------------------------------
# Airflow Providers (إضافات Airflow)
# -----------------------------------------
apache-airflow-providers-postgres==5.12.0
apache-airflow-providers-http==4.12.0

# -----------------------------------------
# Database Drivers (قواعد البيانات)
# -----------------------------------------
psycopg2-binary==2.9.9
SQLAlchemy==1.4.53

# -----------------------------------------
# Data Processing (معالجة البيانات)
# -----------------------------------------
numpy==1.26.4
pandas==2.2.3
pytz==2024.2

# -----------------------------------------
# Machine Learning (التعلم الآلي)
# -----------------------------------------
scikit-learn==1.5.2
xgboost==1.6.2
joblib==1.4.2

# -----------------------------------------
# HTTP & API Requests (طلبات الإنترنت)
# -----------------------------------------
requests==2.32.3
urllib3==2.2.3
certifi==2024.8.30

# -----------------------------------------
# Security & Encryption (الأمان والتشفير)
# -----------------------------------------
pyOpenSSL==24.2.1
cryptography==42.0.8
ndg-httpsclient==0.5.1
pyasn1==0.6.1